rasa/requirements.txt
rasa==3.5.10
requests==2.28.2
python-dotenv==1.0.0